Forward
When a request to a Servlet, it forward the request to another page (or
another servlet). Address on the user's browser is still the linkof the
first page, but the content is created by the forwarded page.

The page was forwarded to, compulsory is a page (or servlet) located on
your web application.