Redirect
When a request from the user to a Servlet (page A), this servlet can
redirect requests to another page (page B), and the end of its mission.
The page is redirected to  may be the page in your application, or may
be a any page.

Unlike Forward. With Redirect you can not use request.setAttribute(..)
to transfer data from page A to page B.