Session
The HttpSession object represents a user session. A user session
contains information about the user across multiple HTTP requests.

When a user enters your site for the first time, the user is given a
unique ID to identify his session by. This ID is typically stored in a
cookie or in a request parameter.